detector/requirements.txt
fastapi
uvicorn[standard]
opencv-python-headless
python-dotenv
numpy
python-multipart